Fixed LinkedCond bug
When a LinkedCond object had many requirements including other LinkedCond objects (number slightly varied), when adding a new requirement, it would sometimes skip one of the related LinkedCond objects from the new results. This resulted in a crash, because the require function would then attempt to set a new main value to that skipped LinkedCond object, but the result dictionary didn't have that object's ID in its keys because it had been skipped.
It seems that this was a bug related to the way limitations were tested for whether they existed in a limitation list. Limitations are tuples, consisting of four elements: an initial string, a tuple of Cond objects, another string and another tuple of numbers. Simply checking if a limitation existed in a list of limitations with "not in" was incorrect, because this compared each element using the "==" operator. For example, the tuples ("", LinkedCond(1,2)) and (", LinkedCond(1,3)) will be equal when checked with the "==" operator.
The fix consists of creating a function that specifically checks if two limitations are the same. For the LinkedCond objects, the "is" operator is used. This ensures that the objects will actually be the same and not just have the same main value.